Move code for displaying a) the initiatives list and b) initiative info sidebar from DirectorySidebarView to a) BaseSidebarView and b) SidebarView
